tests/requirements.txt
qrcode>=7.4
Pillow>=9.0
pyzbar>=0.1.9
